requirements.txt
GitPython==2.1.15
pypinyin==0.42.0